+++ Forge/src/main/java/dev/upcraft/origins/icarae/forge/mixin/compat/origins_fabric/IcarusHelperMixin.java
@@ -5,8 +5,7 @@ import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import dev.cammiescorner.icarus.api.IcarusPlayerValues;
 import dev.cammiescorner.icarus.util.IcarusHelper;
-import dev.upcraft.origins.icarae.forge.compat.origins_fabric.power.WingsPower;
-import io.github.apace100.apoli.component.PowerHolderComponent;
+import dev.upcraft.origins.icarae.forge.compat.origins_fabric.util.IcaraeCompatMixinHooks;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import org.spongepowered.asm.mixin.Mixin;
@@ -20,24 +19,17 @@ public class IcarusHelperMixin {
 
     @ModifyReturnValue(method = "getConfigValues", at = @At("RETURN"))
     private static IcarusPlayerValues injectPowerConfigValues(IcarusPlayerValues original, LivingEntity entity) {
-        var list = PowerHolderComponent.getPowers(entity, WingsPower.class);
-        if(!list.isEmpty()) {
-            var power = list.get(0);
-            power.updateFallback(original);
-            return power;
-        }
-
-        return original;
+        return IcaraeCompatMixinHooks.getConfigOverride(original, entity);
     }
 
     @WrapOperation(method = "hasWings", at = @At(value = "INVOKE", target = "Ljava/util/function/Predicate;test(Ljava/lang/Object;)Z"))
     private static boolean originHasWings(Predicate<LivingEntity> instance, Object entity, Operation<Boolean> original) {
-        return PowerHolderComponent.hasPower((LivingEntity) entity, WingsPower.class) || original.call(instance, entity);
+        return IcaraeCompatMixinHooks.hasWingsOverride((LivingEntity) entity) || original.call(instance, entity);
     }
 
     @WrapOperation(method = "getEquippedWings", at = @At(value = "INVOKE", target = "Ljava/util/function/Function;apply(Ljava/lang/Object;)Ljava/lang/Object;"))
     private static Object originGetWings(Function<LivingEntity, ItemStack> instance, Object entity, Operation<ItemStack> original) {
-        if (PowerHolderComponent.hasPower((LivingEntity) entity, WingsPower.class)) {
+        if (IcaraeCompatMixinHooks.hasWingsOverride((LivingEntity) entity)) {
             // null is special case
             return null;
         }
